refactoring(integrations): useIntegrations hook to isolate all logic out of the component
